requirements.txt
asposestorage==1.0.2
beautifulsoup4==4.13.3
callbacks==0.3.0
datasets==2.18.0
diffusers==0.32.2
faiss_cpu==1.8.0.post1
ftfy==6.3.1
huggingface_hub==0.26.3
matplotlib==3.7.4
numpy==1.24.4
pandas==2.0.3
Pillow==11.1.0
pipeline_utils==0.0.2
pytorch_fid==0.3.0
scheduling_utils==0.2.3
scikit_learn==1.3.2
scipy==1.3.3
skimage==0.0
tqdm==4.66.1
transformers==4.46.3
accelerate==1.0.1
sentencepiece==0.1.99